Renamed Ingredients component to MultiInput, added method to newRecipe state in AddRecipe screen, removed required from prep_time and cook_time input fields in AddRecipe screen, added MultiInput component for method to AddRecipe screen
